Standize demo resource names
