+++ server/config/demo/demo_repository.go
@@ -1,4 +1,4 @@
-package demoresource
+package demo
 
 import (
 	"context"
@@ -10,59 +10,6 @@ import (
 	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/resourcemanager"
 )
-
-type DemoType string
-
-const (
-	DemoTypePokeshop           DemoType = "pokeshop"
-	DemoTypeOpentelemetryStore DemoType = "otelstore"
-)
-
-const (
-	ResourceName       = "Demo"
-	ResourceNamePlural = "Demos"
-)
-
-var Operations = []resourcemanager.Operation{
-	resourcemanager.OperationCreate,
-	resourcemanager.OperationDelete,
-	resourcemanager.OperationGet,
-	resourcemanager.OperationList,
-	resourcemanager.OperationUpdate,
-}
-
-type Demo struct {
-	ID                 id.ID                   `json:"id"`
-	Name               string                  `json:"name"`
-	Type               DemoType                `json:"type"`
-	Enabled            bool                    `json:"enabled"`
-	Pokeshop           *PokeshopDemo           `json:"pokeshop,omitempty"`
-	OpenTelemetryStore *OpenTelemetryStoreDemo `json:"opentelemetryStore,omitempty"`
-}
-
-func (d Demo) HasID() bool {
-	return d.ID != ""
-}
-
-func (d Demo) GetID() id.ID {
-	return d.ID
-}
-
-func (d Demo) Validate() error {
-	return nil
-}
-
-type PokeshopDemo struct {
-	HTTPEndpoint string `json:"httpEndpoint,omitempty"`
-	GRPCEndpoint string `json:"grpcEndpoint,omitempty"`
-}
-
-type OpenTelemetryStoreDemo struct {
-	FrontendEndpoint       string `json:"frontendEndpoint,omitempty"`
-	ProductCatalogEndpoint string `json:"productCatalogEndpoint,omitempty"`
-	CartEndpoint           string `json:"cartEndpoint,omitempty"`
-	CheckoutEndpoint       string `json:"checkoutEndpoint,omitempty"`
-}
 
 type Repository struct {
 	db *sql.DB
